move these things out of the callback
